# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: TourFlow AI
  description: 'TourFlow AI API contract — generated from the FastAPI implementation
    (source of truth: backend/app). Regenerate: python -c "from app.main import app;
    ..."'
  version: 1.0.0
paths:
...
  /api/v1/trips/{trip_id}/itinerary:
    get:
      tags:
      - itinerary
      summary: List Items
      operationId: list_items_api_v1_trips__trip_id__itinerary_get
      security:
      - HTTPBearer: []
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Trip Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ItineraryItemOut'
                title: Response List Items Api V1 Trips  Trip Id  Itinerary Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - itinerary
      summary: Create Item
      operationId: create_item_api_v1_trips__trip_id__itinerary_post
      security:
      - HTTPBearer: []
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Trip Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItineraryItemCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItineraryItemOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/itinerary/{item_id}:
    patch:
      tags:
      - itinerary
      summary: Update Item
      operationId: update_item_api_v1_itinerary__item_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Item Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItineraryItemUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItineraryItemOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - itinerary
      summary: Delete Item
      operationId: delete_item_api_v1_itinerary__item_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Item Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/trips/{trip_id}/dependencies:
    get:
      tags:
      - itinerary
      summary: List Dependencies
      operationId: list_dependencies_api_v1_trips__trip_id__dependencies_get
      security:
      - HTTPBearer: []
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Trip Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DependencyOut'
                title: Response List Dependencies Api V1 Trips  Trip Id  Dependencies
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - itinerary
      summary: Create Dependency
      operationId: create_dependency_api_v1_trips__trip_id__dependencies_post
      security:
      - HTTPBearer: []
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Trip Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DependencyCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DependencyOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/dependencies/{dep_id}:
    delete:
      tags:
      - itinerary
      summary: Delete Dependency
      operationId: delete_dependency_api_v1_dependencies__dep_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: dep_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Dep Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/trips/{trip_id}/disruptions:
    post:
      tags:
      - disruptions
      summary: Create Disruption
      operationId: create_disruption_api_v1_trips__trip_id__disruptions_post
      security:
      - HTTPBearer: []
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Trip Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DisruptionCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DisruptionOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - disruptions
      summary: List Disruptions
      operationId: list_disruptions_api_v1_trips__trip_id__disruptions_get
      security:
      - HTTPBearer: []
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Trip Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DisruptionOut'
                title: Response List Disruptions Api V1 Trips  Trip Id  Disruptions
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/disruptions/{disruption_id}/impact:
    post:
      tags:
      - disruptions
      summary: Run Impact
      operationId: run_impact_api_v1_disruptions__disruption_id__impact_post
      security:
      - HTTPBearer: []
      parameters:
      - name: disruption_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Disruption Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImpactOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/disruptions/{disruption_id}/recovery-options:
    post:
      tags:
      - disruptions
      summary: Create Recovery Options
      operationId: create_recovery_options_api_v1_disruptions__disruption_id__recovery_options_post
      security:
      - HTTPBearer: []
      parameters:
      - name: disruption_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Disruption Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecoveryOptionOut'
                title: Response Create Recovery Options Api V1 Disruptions  Disruption
                  Id  Recovery Options Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - disruptions
      summary: List Recovery Options
      operationId: list_recovery_options_api_v1_disruptions__disruption_id__recovery_options_get
      security:
      - HTTPBearer: []
      parameters:
      - name: disruption_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Disruption Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecoveryOptionOut'
                title: Response List Recovery Options Api V1 Disruptions  Disruption
                  Id  Recovery Options Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/recovery-options/{option_id}/select:
    post:
      tags:
      - disruptions
      summary: Select Option
      operationId: select_option_api_v1_recovery_options__option_id__select_post
      security:
      - HTTPBearer: []
      parameters:
      - name: option_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Option Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecoveryOptionOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    DependencyCreate:
      properties:
        source_item_id:
          type: string
          format: uuid
          title: Source Item Id
        target_item_id:
          type: string
          format: uuid
          title: Target Item Id
        dependency_type:
          type: string
          pattern: ^(SEQUENTIAL|SPATIAL)$
          title: Dependency Type
          default: SEQUENTIAL
        minimum_required_buffer_minutes:
          type: integer
          maximum: 1440.0
          minimum: 0.0
          title: Minimum Required Buffer Minutes
          default: 30
      type: object
      required:
      - source_item_id
      - target_item_id
      title: DependencyCreate
    DependencyOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        trip_id:
          type: string
          format: uuid
          title: Trip Id
        source_item_id:
          type: string
          format: uuid
          title: Source Item Id
        target_item_id:
          type: string
          format: uuid
          title: Target Item Id
        dependency_type:
          type: string
          title: Dependency Type
        minimum_required_buffer_minutes:
          type: integer
          title: Minimum Required Buffer Minutes
      type: object
      required:
      - id
      - trip_id
      - source_item_id
      - target_item_id
      - dependency_type
      - minimum_required_buffer_minutes
      title: DependencyOut
...
    DisruptionCreate:
      properties:
        type:
          $ref: '#/components/schemas/DisruptionType'
        source_item_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Source Item Id
        new_start_time:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: New Start Time
        delay_minutes:
          type: integer
          maximum: 10080.0
          minimum: 0.0
          title: Delay Minutes
          default: 0
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
      type: object
      required:
      - type
      title: DisruptionCreate
    DisruptionOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        trip_id:
          type: string
          format: uuid
          title: Trip Id
        type:
          $ref: '#/components/schemas/DisruptionType'
        source_item_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Source Item Id
        original_start_time:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Original Start Time
        new_start_time:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: New Start Time
        delay_minutes:
          type: integer
          title: Delay Minutes
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
        status:
          $ref: '#/components/schemas/DisruptionStatus'
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - trip_id
      - type
      - delay_minutes
      - status
      - created_at
      title: DisruptionOut
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ImpactOut:
      properties:
        evaluation_id:
          type: string
          format: uuid
          title: Evaluation Id
        feasible:
          type: boolean
          title: Feasible
        affected_items:
          additionalProperties:
            type: string
          type: object
          title: Affected Items
        broken_items:
          items:
            type: string
          type: array
          title: Broken Items
        at_risk_items:
          items:
            type: string
          type: array
          title: At Risk Items
        violations:
          items:
            type: object
          type: array
          title: Violations
        warnings:
          items:
            type: object
          type: array
          title: Warnings
        explanation:
          type: string
          title: Explanation
      type: object
      required:
      - evaluation_id
      - feasible
      - affected_items
      - broken_items
      - at_risk_items
      - violations
      - warnings
      - explanation
      title: ImpactOut
    ItemStatus:
      type: string
      enum:
      - PLANNED
      - CONFIRMED
      - AT_RISK
      - BROKEN
      - CANCELLED
      - COMPLETED
      - REBOOKED
      title: ItemStatus
    ItemType:
      type: string
      enum:
      - FLIGHT
      - TRAIN
      - BUS
      - TRANSFER
      - HOTEL
      - ACTIVITY
      - EVENT
      - RESTAURANT
      - OTHER
      title: ItemType
    ItineraryItemCreate:
      properties:
        type:
          $ref: '#/components/schemas/ItemType'
        title:
          type: string
          maxLength: 250
          minLength: 2
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        location:
          anyOf:
          - type: string
            maxLength: 250
          - type: 'null'
          title: Location
        start_time:
          type: string
          format: date-time
          title: Start Time
        end_time:
          type: string
          format: date-time
          title: End Time
        duration_minutes:
          anyOf:
          - type: integer
            minimum: 0.0
          - type: 'null'
          title: Duration Minutes
        cost:
          anyOf:
          - type: number
            minimum: 0.0
          - type: 'null'
          title: Cost
        currency:
          type: string
          maxLength: 3
          minLength: 3
          title: Currency
          default: INR
        is_fixed:
          type: boolean
          title: Is Fixed
          default: false
        vendor_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Vendor Id
        sequence_order:
          type: integer
          title: Sequence Order
          default: 0
      type: object
      required:
      - type
      - title
      - start_time
      - end_time
      title: ItineraryItemCreate
    ItineraryItemOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        trip_id:
          type: string
          format: uuid
          title: Trip Id
        type:
          $ref: '#/components/schemas/ItemType'
        title:
          type: string
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        location:
          anyOf:
          - type: string
          - type: 'null'
          title: Location
        start_time:
          type: string
          format: date-time
          title: Start Time
        end_time:
          type: string
          format: date-time
          title: End Time
        duration_minutes:
          anyOf:
          - type: integer
          - type: 'null'
          title: Duration Minutes
        cost:
          anyOf:
          - type: number
          - type: 'null'
          title: Cost
        currency:
          type: string
          title: Currency
        status:
          $ref: '#/components/schemas/ItemStatus'
        booked_status:
          $ref: '#/components/schemas/BookingStatus'
        is_fixed:
          type: boolean
          title: Is Fixed
        vendor_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Vendor Id
        booking_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Booking Id
        sequence_order:
          type: integer
          title: Sequence Order
      type: object
      required:
      - id
      - trip_id
      - type
      - title
      - start_time
      - end_time
      - currency
      - status
      - booked_status
      - is_fixed
      - sequence_order
      title: ItineraryItemOut
    ItineraryItemUpdate:
      properties:
        type:
          anyOf:
          - $ref: '#/components/schemas/ItemType'
          - type: 'null'
        title:
          anyOf:
          - type: string
            maxLength: 250
            minLength: 2
          - type: 'null'
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        location:
          anyOf:
          - type: string
            maxLength: 250
          - type: 'null'
          title: Location
        start_time:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Start Time
        end_time:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: End Time
        duration_minutes:
          anyOf:
          - type: integer
            minimum: 0.0
          - type: 'null'
          title: Duration Minutes
        cost:
          anyOf:
          - type: number
            minimum: 0.0
          - type: 'null'
          title: Cost
        status:
          anyOf:
          - $ref: '#/components/schemas/ItemStatus'
          - type: 'null'
        is_fixed:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Fixed
        vendor_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Vendor Id
        sequence_order:
          anyOf:
          - type: integer
          - type: 'null'
          title: Sequence Order
      type: object
      title: ItineraryItemUpdate
...
    RecoveryOptionOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        trip_id:
          type: string
          format: uuid
          title: Trip Id
        disruption_id:
          type: string
          format: uuid
          title: Disruption Id
        action:
          $ref: '#/components/schemas/RecoveryAction'
        title:
          type: string
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        estimated_cost_delta:
          type: number
          title: Estimated Cost Delta
        experience_impact:
          anyOf:
          - type: string
          - type: 'null'
          title: Experience Impact
        feasibility:
          type: boolean
          title: Feasibility
        ai_rank:
          anyOf:
          - type: integer
          - type: 'null'
          title: Ai Rank
        ai_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Ai Reason
        affected_items:
          items:
            type: string
          type: array
          title: Affected Items
        changes:
          type: object
          title: Changes
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
        selected:
          type: boolean
          title: Selected
      type: object
      required:
      - id
      - trip_id
      - disruption_id
      - action
      - title
      - estimated_cost_delta
      - feasibility
      - affected_items
      - changes
      - selected
      title: RecoveryOptionOut
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer